fix ModifierCoefficient type

+++ Assets/Scripts/Components/SpeedModifierComponent.cs
@@ -2,6 +2,6 @@ using Unity.Entities;
 
 public struct SpeedModifierComponent : IEnableableComponent, IComponentData
 {
-    public int ModifierCoefficient;
+    public float ModifierCoefficient;
     public float Timeout;
 }
